pyproject.toml
[tool.poetry]
name = "hangugeoreul-gongbuhaeyo"
version = "0.1.0"
description = "Korean Language Learn Helper."
authors = ["Rafail Medzhidov <rafayt323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

django = { version = "^4.2", extras = ["argon2"] }
django-split-settings = "^1.2"
django-axes = "^6.1"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.18"
django-stubs-ext = "^4.2"

psycopg2-binary = "^2.9"
gunicorn = "^21.2"
python-decouple = "^3.8"
structlog = "^23.2"
djangorestframework = "^3.14.0"
jsonschema = "^4.19.2"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2"
django-querycount = "^0.8"
django-migration-linter = "^5.0"
django-extra-checks = "^0.13"
nplusone = "^1.0"

wemake-python-styleguide = "^0.18"
flake8-pytest-style = "^1.7"
flake8-django = "^1.4"
flake8-logging-format = "^0.9"

pytest = "^7.4"
pytest-django = "^4.5"
django-test-migrations = "^1.3"

pyright = "^1.1.333"

django-stubs = { version = "^4.2.6" }
pre-commit = "^3.5.0"

[tool.pyright]
venvPath = '.'
venv = '.venv'
pythonVersion = '3.11'
verboseOutput = true
ignore = ['**/server/settings/*']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
